Destroying bullets when hit
Made a system to (properly) spawn the bullets from both the player and the enemy  - (but have to take care of them getting shifted when spawned). Also added a mechanism to get them to disappear when they collide with an enemy or player. Still have to correct bullet-bullet collisions and set player and enemies to rigidbodies.